catshef/requirements.txt
Babel==2.3.4
coverage==4.2
defusedxml==0.4.1
Django==1.9.8
django-allauth==0.26.1
django-bootstrap3==7.0.1
django-nose==1.4.4
django-phonenumber-field==1.1.0
nose==1.3.7
oauthlib==1.1.2
phonenumberslite==7.5.2
Pillow==3.3.0
python3-openid==3.0.10
pytz==2016.6.1
requests==2.11.0
requests-oauthlib==0.6.2
selenium==2.53.6
sorl-thumbnail==12.3
